add line, fill, circle, symbol query tests
